Single data set with a custom color should not create legends for each category
When creating a bar chart with a single data set and setting custom colors each category gets a legend-entry. See #285 for last time that happened.

I guess the idea is that if you pass in an array of colors to a single data set, that means that you want each category to have a different color.

I think it would make more sense if creating legend entries for categories were an option that could be set.
This solution instead add a check to avoid trying to color each category, if only a single color is passed.